+++ src/modules/mo-helm/data-transfer-objects/helm-entry-status.dto.ts
@@ -1,0 +1,14 @@
+import { Expose, Type } from 'class-transformer';
+import { HelmEntryDto } from './helm-entry.dto';
+
+export class HelmEntryStatusDto {
+  @Type(() => HelmEntryDto)
+  @Expose()
+  entry: HelmEntryDto;
+
+  @Expose()
+  status: 'success' | 'error';
+
+  @Expose()
+  errorMessage?: string;
+}