+++ src/modules/mo-helm/data-transfer-objects/release/cluster-helm-release-list-request.dto.ts
@@ -1,0 +1,10 @@
+import { IsOptional, IsString } from 'class-validator';
+import { Expose, Transform } from 'class-transformer';
+
+export class ClusterHelmReleaseListRequestDto {
+  @IsOptional()
+  @Transform(({ value }) => value ?? '')
+  @IsString()
+  @Expose()
+  namespace: string;
+}