+++ src/modules/mo-helm/data-transfer-objects/release/cluster-helm-release-status-info.dto.ts
@@ -1,7 +1,7 @@
 import { Expose, Transform } from 'class-transformer';
 import moment from 'moment';
 
-export class HelmReleaseStatusInfoDto {
+export class ClusterHelmReleaseStatusInfoDto {
   @Expose()
   name: string;