+++ src/modules/mo-helm/data-transfer-objects/repo/cluster-helm-repo-patch-request.dto.ts
@@ -1,0 +1,10 @@
+import { IsNotEmpty, IsString } from 'class-validator';
+import { Expose } from 'class-transformer';
+import { ClusterHelmRepoAddRequestDto } from './cluster-helm-repo-add-request.dto';
+
+export class ClusterHelmRepoPatchRequestDto extends ClusterHelmRepoAddRequestDto {
+  @IsNotEmpty()
+  @IsString()
+  @Expose()
+  newName: string;
+}