+++ src/modules/mo-helm/data-transfer-objects/repo/cluster-helm-repo-update-response.dto.ts
@@ -1,0 +1,10 @@
+import { Expose, Transform } from 'class-transformer';
+import { K8sMessageResponseBaseDto } from '../../../mo-kubernetes/data-transfer-objects/k8s-message-response-base.dto';
+import { MoUtils } from '@mogenius/js-utils';
+import { HelmEntryStatusDto } from '../helm-entry-status.dto';
+
+export class ClusterHelmRepoUpdateResponseDto extends K8sMessageResponseBaseDto<HelmEntryStatusDto[]> {
+  @Transform(({ value }) => MoUtils.transformToDtoList(HelmEntryStatusDto, value))
+  @Expose()
+  data?: HelmEntryStatusDto[];
+}