+++ src/modules/mo-kubernetes/data-transfer-objects/index.ts
@@ -4,6 +4,8 @@ export * from './k8s-manager-job';
 export * from './k8s-manager-traffic';
 export * from './k8s-resources';
 
+export * from './k8s-message-response-base.dto';
+export * from './k8s-message-response.dto';
 export * from './kubernetes-event.dto';
 export * from './kubernetes-event-involved-object.dto';
 export * from './kubernetes-event-metadata.dto';