+++ src/modules/mo-kubernetes/data-transfer-objects/k8s-message-response.dto.ts
@@ -1,0 +1,7 @@
+import { Expose } from 'class-transformer';
+import { K8sMessageResponseBaseDto } from './k8s-message-response-base.dto';
+
+export class K8sMessageResponseDto extends K8sMessageResponseBaseDto<string> {
+  @Expose()
+  data?: string;
+}