+++ src/modules/mo-project-dto/data-transfer-objects/external-secret-storage/store/external-secret-store.dto.ts
@@ -1,6 +1,18 @@
 import { Expose, Transform } from 'class-transformer';
+import moment from 'moment';
+import { isNotEmpty } from 'class-validator';
+import { MoUtils, StripTags } from '@mogenius/js-utils';
 
 export class ExternalSecretStoreDto {
+  @Transform(({ value }) => (value && value !== 'undefined' && value !== 'null' ? moment(value).toDate() : value))
+  @Expose()
+  createdAt: Date;
+
+  @StripTags()
+  @Transform(({ value, obj }) => (value && isNotEmpty(value) ? MoUtils.alphaNumeric(value, obj.name) : obj.name))
+  @Expose()
+  displayName: string;
+
   @Expose()
   name: string;
 
@@ -21,4 +33,28 @@ export class ExternalSecretStoreDto {
   @Transform(({ value, obj }) => value ?? obj['vault-url'], { toClassOnly: true })
   @Expose()
   vaultURL: string;
+
+  @Transform(({ value }) => {
+    switch (value) {
+      case 'True':
+      case true:
+        return true;
+      case 'False':
+      case false:
+        return false;
+      default:
+        return 'Unknown';
+    }
+  })
+  @Expose()
+  status: boolean | 'Unknown';
+
+  @Expose()
+  message: string;
+
+  @Expose()
+  reason: string;
+
+  @Expose()
+  type: 'Deleted' | 'Ready';
 }