+++ src/modules/mo-project-dto/data-transfer-objects/project/project-notification-patch.request.dto.ts
@@ -1,0 +1,13 @@
+import { Expose, Type } from 'class-transformer';
+import { IsString } from 'class-validator';
+import { ProjectNotificationSettingsBaseDto } from './project-notification-settings.base.dto';
+
+export class ProjectNotificationPatchRequestDto {
+  @IsString()
+  @Expose()
+  id: string;
+
+  @Type(() => ProjectNotificationSettingsBaseDto)
+  @Expose()
+  notificationSettings: ProjectNotificationSettingsBaseDto[];
+}