+++ src/modules/mo-project-dto/enums/project-namespace-service-envvar-type.enum.ts
@@ -5,6 +5,7 @@ export enum ProjectNamespaceServiceEnvVarTypeEnum {
   // VOLUME_MOUNT_SEED = 'VOLUME_MOUNT_SEED',
   // CHANGE_OWNER = 'CHANGE_OWNER',
   HOSTNAME = 'HOSTNAME'
+  // EXTERNAL_SECRET_STORE = 'EXTERNAL_SECRET_STORE'
 }
 
 export const ProjectNamespaceServiceEnvVarTypeEnum_ALL = [
@@ -14,4 +15,5 @@ export const ProjectNamespaceServiceEnvVarTypeEnum_ALL = [
   // ProjectNamespaceServiceEnvVarTypeEnum.VOLUME_MOUNT_SEED,
   // ProjectNamespaceServiceEnvVarTypeEnum.CHANGE_OWNER,
   ProjectNamespaceServiceEnvVarTypeEnum.HOSTNAME
+  // ProjectNamespaceServiceEnvVarTypeEnum.EXTERNAL_SECRET_STORE
 ];